+++ genomics_data_index/storage/service/__init__.py
@@ -1,5 +1,5 @@
-from pathlib import Path
 import logging
+from pathlib import Path
 
 import pandas as pd
 from sqlalchemy import create_engine