+++ genomics_data_index/test/integration/api/query/test_GenomicsDataIndex_samples_query.py
@@ -1665,7 +1665,7 @@ def test_join_kmer_tree(prebuilt_tree: Tree, loaded_database_connection: DataInd
     assert {'SampleA', 'SampleB', 'SampleC'} == set(query_result.tree.get_leaf_names())
 
 
-def test_summary_features_all(loaded_database_connection: DataIndexConnection):
+def test_summary_features_kindmutations(loaded_database_connection: DataIndexConnection):
     dfA = pd.read_csv(snippy_all_dataframes['SampleA'], sep='\t')
     dfB = pd.read_csv(snippy_all_dataframes['SampleB'], sep='\t')
     dfC = pd.read_csv(snippy_all_dataframes['SampleC'], sep='\t')
@@ -1677,13 +1677,151 @@ def test_summary_features_all(loaded_database_connection: DataIndexConnection):
         'Insertion': 'first',
         'Mutation': 'count',
     }).rename(columns={'Mutation': 'Count'}).sort_index()
+    expected_df['Total'] = 9
+    expected_df['Percent'] = 100 * (expected_df['Count'] / expected_df['Total'])
 
     mutations_df = query(loaded_database_connection).summary_features()
     mutations_df = mutations_df.sort_index()
 
     assert len(expected_df) == len(mutations_df)
+    assert list(expected_df.columns) == list(mutations_df.columns)
     assert list(expected_df.index) == list(mutations_df.index)
     assert list(expected_df['Count']) == list(mutations_df['Count'])
+    assert list(expected_df['Total']) == list(mutations_df['Total'])
+    assert math.isclose(100 * (2 / 9), mutations_df.loc['reference:619:G:C', 'Percent'])
+
+
+def test_summary_features_kindmutations_unique(loaded_database_connection: DataIndexConnection):
+    dfA = pd.read_csv(snippy_all_dataframes['SampleA'], sep='\t')
+    dfB = pd.read_csv(snippy_all_dataframes['SampleB'], sep='\t')
+    dfC = pd.read_csv(snippy_all_dataframes['SampleC'], sep='\t')
+
+    # Unique to A
+    expected_df = dfA
+    expected_df = expected_df.groupby('Mutation').agg({
+        'Sequence': 'first',
+        'Position': 'first',
+        'Deletion': 'first',
+        'Insertion': 'first',
+        'Mutation': 'count',
+    }).rename(columns={'Mutation': 'Count'}).sort_index()
+    expected_df['Total'] = 1
+    expected_df['Percent'] = 100 * (expected_df['Count'] / expected_df['Total'])
+
+    q = query(loaded_database_connection)
+
+    mutations_df = q.isa('SampleA').summary_features(selection='unique')
+    mutations_df = mutations_df.sort_index()
+
+    assert len(expected_df) == len(mutations_df)
+    assert 46 == len(mutations_df)  # Check length against independently generated length
+    assert list(expected_df.index) == list(mutations_df.index)
+    assert list(expected_df['Count']) == list(mutations_df['Count'])
+    assert list(expected_df['Total']) == list(mutations_df['Total'])
+    assert math.isclose(100 * (1 / 1), mutations_df.loc['reference:3656:CATT:C', 'Percent'])
+
+    # Unique to B
+    dfAC = pd.concat([dfA, dfC])
+    expected_df = dfB[~dfB['Mutation'].isin(list(dfAC['Mutation']))]
+    expected_df = expected_df.groupby('Mutation').agg({
+        'Sequence': 'first',
+        'Position': 'first',
+        'Deletion': 'first',
+        'Insertion': 'first',
+        'Mutation': 'count',
+    }).rename(columns={'Mutation': 'Count'}).sort_index()
+    expected_df['Total'] = 1
+    expected_df['Percent'] = 100 * (expected_df['Count'] / expected_df['Total'])
+
+    mutations_df = q.isa('SampleB').summary_features(selection='unique')
+    mutations_df = mutations_df.sort_index()
+
+    assert len(expected_df) == len(mutations_df)
+    assert list(expected_df.index) == list(mutations_df.index)
+    assert list(expected_df['Count']) == list(mutations_df['Count'])
+    assert list(expected_df['Total']) == list(mutations_df['Total'])
+    assert math.isclose(100 * (1 / 1), mutations_df.loc['reference:349:AAGT:A', 'Percent'])
+
+    # Unique to C
+    dfAB = pd.concat([dfA, dfB])
+    expected_df = dfC[~dfC['Mutation'].isin(list(dfAB['Mutation']))]
+    expected_df = expected_df.groupby('Mutation').agg({
+        'Sequence': 'first',
+        'Position': 'first',
+        'Deletion': 'first',
+        'Insertion': 'first',
+        'Mutation': 'count',
+    }).rename(columns={'Mutation': 'Count'}).sort_index()
+    expected_df['Total'] = 1
+    expected_df['Percent'] = 100 * (expected_df['Count'] / expected_df['Total'])
+
+    mutations_df = q.isa('SampleC').summary_features(selection='unique')
+    mutations_df = mutations_df.sort_index()
+
+    assert len(expected_df) == len(mutations_df)
+    assert list(expected_df.index) == list(mutations_df.index)
+    assert list(expected_df['Count']) == list(mutations_df['Count'])
+    assert list(expected_df['Total']) == list(mutations_df['Total'])
+    assert math.isclose(100 * (1 / 1), mutations_df.loc['reference:866:GCCAGATCC:G', 'Percent'])
+
+    # Unique to BC
+    dfBC = pd.concat([dfB, dfC])
+    expected_df = dfBC[~dfBC['Mutation'].isin(list(dfA['Mutation']))]
+    expected_df = expected_df.groupby('Mutation').agg({
+        'Sequence': 'first',
+        'Position': 'first',
+        'Deletion': 'first',
+        'Insertion': 'first',
+        'Mutation': 'count',
+    }).rename(columns={'Mutation': 'Count'}).sort_index()
+    expected_df['Total'] = 2
+    expected_df['Percent'] = 100 * (expected_df['Count'] / expected_df['Total'])
+
+    mutations_df = q.isin(['SampleB', 'SampleC']).summary_features(selection='unique')
+    mutations_df = mutations_df.sort_index()
+
+    assert len(expected_df) == len(mutations_df)
+    assert 66 == len(mutations_df)  # Check length against independently generated length
+    assert list(expected_df.index) == list(mutations_df.index)
+    assert list(expected_df['Count']) == list(mutations_df['Count'])
+    assert list(expected_df['Total']) == list(mutations_df['Total'])
+    assert math.isclose(100 * (2 / 2), mutations_df.loc['reference:619:G:C', 'Percent'])
+    assert math.isclose(100 * (1 / 2), mutations_df.loc['reference:866:GCCAGATCC:G', 'Percent'])
+    assert math.isclose(100 * (1 / 2), mutations_df.loc['reference:349:AAGT:A', 'Percent'])
+
+    # Unique to ABC (all with mutations)
+    dfABC = pd.concat([dfA, dfB, dfC])
+    expected_df = dfABC
+    expected_df = expected_df.groupby('Mutation').agg({
+        'Sequence': 'first',
+        'Position': 'first',
+        'Deletion': 'first',
+        'Insertion': 'first',
+        'Mutation': 'count',
+    }).rename(columns={'Mutation': 'Count'}).sort_index()
+    expected_df['Total'] = 3
+    expected_df['Percent'] = 100 * (expected_df['Count'] / expected_df['Total'])
+
+    mutations_df = q.isin(['SampleA', 'SampleB', 'SampleC']).summary_features(selection='unique')
+    mutations_df = mutations_df.sort_index()
+
+    assert len(expected_df) == len(mutations_df)
+    assert 112 == len(mutations_df)  # Check length against independently generated length
+    assert list(expected_df.index) == list(mutations_df.index)
+    assert list(expected_df['Count']) == list(mutations_df['Count'])
+    assert list(expected_df['Total']) == list(mutations_df['Total'])
+    assert math.isclose(100 * (2 / 3), mutations_df.loc['reference:619:G:C', 'Percent'])
+    assert math.isclose(100 * (1 / 3), mutations_df.loc['reference:866:GCCAGATCC:G', 'Percent'])
+    assert math.isclose(100 * (1 / 3), mutations_df.loc['reference:349:AAGT:A', 'Percent'])
+    assert math.isclose(100 * (1 / 3), mutations_df.loc['reference:3656:CATT:C', 'Percent'])
+
+    # Unique to None
+    mutations_df = q.isin([]).summary_features(selection='unique')
+    mutations_df = mutations_df.sort_index()
+
+    assert 0 == len(mutations_df)
+    assert ['Sequence', 'Position', 'Deletion', 'Insertion',
+            'Count', 'Total', 'Percent'] == list(mutations_df.columns)
 
 
 def test_summary_features_two(loaded_database_connection: DataIndexConnection):
@@ -1697,6 +1835,8 @@ def test_summary_features_two(loaded_database_connection: DataIndexConnection):
         'Insertion': 'first',
         'Mutation': 'count',
     }).rename(columns={'Mutation': 'Count'}).sort_index()
+    expected_df['Total'] = 2
+    expected_df['Percent'] = 100 * (expected_df['Count'] / expected_df['Total'])
 
     mutations_df = query(loaded_database_connection).hasa(
         'reference:839:C:G', kind='mutation').summary_features()
@@ -1705,6 +1845,8 @@ def test_summary_features_two(loaded_database_connection: DataIndexConnection):
     assert len(expected_df) == len(mutations_df)
     assert list(expected_df.index) == list(mutations_df.index)
     assert list(expected_df['Count']) == list(mutations_df['Count'])
+    assert list(expected_df['Total']) == list(mutations_df['Total'])
+    assert math.isclose(100 * (2 / 2), mutations_df.loc['reference:839:C:G', 'Percent'])
 
 
 def test_tofeaturesset_all(loaded_database_only_snippy: DataIndexConnection):